io cleanup
- Create io directory
- Seperated each io class into its own file under the io directory
- Created HDF5Reader and HDF5Writer, generic classes to handle the
CHECLabPy file type
- Updated DL1Reader to be a subclass of HDF5Reader
- Deprecated storing of monitor information into the dl1.h5 file
- Updated SimtelReader
- Updated the extract_dl1.py and extract_dl1_simtel.py scripts to use
new io
- Updated CHECLabPy version to 2.0.0
