+++ CHECLabPy/core/io/simtel.py
@@ -1,13 +1,14 @@
 from ctapipe.calib import HESSIOR1Calibrator
-from ctapipe.io import HESSIOEventSource, EventSeeker
+from ctapipe.io import SimTelEventSource, EventSeeker
 from target_calib import CameraConfiguration
 from CHECLabPy.utils.mapping import get_clp_mapping_from_tc_mapping
+import pandas as pd
 
 
-class ReaderSimtel:
+class SimtelReader:
     def __init__(self, path, max_events=None):
-        kwargs = dict(input_url=path, max_events=max_events)
-        reader = HESSIOEventSource(**kwargs)
+        self.path = path
+        reader = SimTelEventSource(input_url=path, max_events=max_events)
         self.seeker = EventSeeker(reader)
 
         first_event = self.seeker[0]
@@ -30,7 +31,10 @@ class ReaderSimtel:
 
         self.r1 = HESSIOR1Calibrator()
 
+        self.gps_time = None
         self.mc_true = None
+        self.mc = None
+        self.pointing=None
 
     @property
     def n_events(self):
@@ -39,9 +43,37 @@ class ReaderSimtel:
     def __iter__(self):
         for event in self.seeker:
             self.index = event.count
+            self.gps_time = event.trig.gps_time
+            self.obs_id = event.r0.obs_id
+
             self.r1.calibrate(event)
             waveforms = event.r1.tel[self.tel].waveform[0]
             self.mc_true = event.mc.tel[self.tel].photo_electron_image
+
+            self.mc = dict(
+                iev=self.index,
+                obs_id=self.obs_id,
+                t_cpu=self.t_cpu,
+                energy=event.mc.energy.value,
+                alt=event.mc.alt.value,
+                az=event.mc.az.value,
+                core_x=event.mc.core_x.value,
+                core_y=event.mc.core_y.value,
+                h_first_int=event.mc.h_first_int.value,
+                x_max=event.mc.x_max.value,
+                shower_primary_id = event.mc.shower_primary_id
+            )
+
+            self.pointing = dict(
+                iev=self.index,
+                obs_id=self.obs_id,
+                t_cpu=self.t_cpu,
+                azimuth_raw=event.mc.tel[self.tel].azimuth_raw,
+                altitude_raw=event.mc.tel[self.tel].altitude_raw,
+                azimuth_cor=event.mc.tel[self.tel].azimuth_cor,
+                altitude_cor=event.mc.tel[self.tel].altitude_cor,
+            )
+
             yield waveforms
 
     def __getitem__(self, iev):
@@ -50,3 +82,7 @@ class ReaderSimtel:
         self.r1.calibrate(event)
         waveforms = event.r1.tel[self.tel].waveform[0]
         return waveforms
+
+    @property
+    def t_cpu(self):
+        return pd.to_datetime(self.gps_time.value, unit='s')
