+++ scripts/extract_dl1.py
@@ -1,6 +1,6 @@
 """
 Executable for processing the R1 waveforms, and storing the reduced parameters
-into a HDF5 file, openable as a `pandas.DataFrame`.
+into a HDF5 file, readable as a `pandas.DataFrame`.
 """
 import argparse
 from argparse import ArgumentDefaultsHelpFormatter as Formatter
@@ -8,9 +8,24 @@ import numpy as np
 import pandas as pd
 from tqdm import tqdm
 import json
-from CHECLabPy.core.io import ReaderR1, DL1Writer
+from CHECLabPy.core.io import ReaderR1, HDF5Writer
 from CHECLabPy.core.factory import WaveformReducerFactory
 from CHECLabPy.utils.waveform import BaselineSubtractor
+
+
+class DL1Writer(HDF5Writer):
+    @staticmethod
+    def _prepare_before_append(df):
+        df_float = df.select_dtypes(
+            include=['float']
+        ).apply(pd.to_numeric, downcast='float')
+        df[df_float.columns] = df_float
+        df['iev'] = df['iev'].astype(np.uint32)
+        df['pixel'] = df['pixel'].astype(np.uint32)
+        df['first_cell_id'] = df['first_cell_id'].astype(np.uint16)
+        df['t_tack'] = df['t_tack'].astype(np.uint64)
+        df = df.sort_values(["iev", "pixel"])
+        return df
 
 
 def main():
@@ -20,8 +35,6 @@ def main():
                                      formatter_class=Formatter)
     parser.add_argument('-f', '--files', dest='input_paths', nargs='+',
                         help='path to the TIO r1 run files')
-    parser.add_argument('-m', '--monitor', dest='monitor', action='store',
-                        help='path to the monitor file (OPTIONAL)')
     parser.add_argument('-o', '--output', dest='output_path', action='store',
                         help='path to store the output HDF5 dl1 file '
                              '(OPTIONAL, will be automatically set if '
@@ -55,6 +68,7 @@ def main():
         n_pixels = reader.n_pixels
         n_samples = reader.n_samples
         n_cells = reader.n_cells
+        expectedrows = n_events * n_pixels
         pixel_array = np.arange(n_pixels)
         camera_version = reader.camera_version
         mapping = reader.mapping
@@ -74,9 +88,11 @@ def main():
         input_path = reader.path
         output_path = args.output_path
         if not output_path:
-            output_path = input_path.rsplit('_r1', 1)[0] + "_dl1.h5"
+            output_path = (
+                input_path.replace('_r1', '_dl1').replace('.tio', '.h5')
+            )
 
-        with DL1Writer(output_path, n_events*n_pixels, args.monitor) as writer:
+        with DL1Writer(output_path) as writer:
             t_cpu = 0
             start_time = 0
             desc = "Processing events"
@@ -84,12 +100,7 @@ def main():
                 iev = reader.index
 
                 t_tack = reader.current_tack
-                t_cpu_sec = reader.current_cpu_s
-                t_cpu_ns = reader.current_cpu_ns
-                t_cpu = pd.to_datetime(
-                    np.int64(t_cpu_sec * 1E9) + np.int64(t_cpu_ns),
-                    unit='ns'
-                )
+                t_cpu = reader.t_cpu
                 fci = reader.first_cell_ids
 
                 if not start_time:
@@ -100,7 +111,7 @@ def main():
 
                 params = reducer.process(waveforms_bs)
 
-                df_ev = pd.DataFrame(dict(
+                df = pd.DataFrame(dict(
                     iev=iev,
                     pixel=pixel_array,
                     first_cell_id=fci,
@@ -109,7 +120,7 @@ def main():
                     baseline_subtracted=bs,
                     **params
                 ))
-                writer.append_event(df_ev)
+                writer.append(df, key='data', expectedrows=expectedrows)
 
             sn_dict = {}
             for tm in range(n_modules):
@@ -129,11 +140,12 @@ def main():
                 camera_version=camera_version,
                 reducer=reducer.__class__.__name__,
                 configuration=config_string,
-                **sn_dict
             )
 
-            writer.add_metadata(**metadata)
             writer.add_mapping(mapping)
+            writer.add_metadata(name='metadata', **metadata)
+            writer.add_metadata(name='sn', **sn_dict)
+            writer.add_metadata(name='config', **config)
 
 
 if __name__ == '__main__':
